- fix exam component
